While loops replaces else if

+++ ADA/src/app/Point.java
@@ -11,6 +11,6 @@ public class Point
     }
     public String toString()
     {
-        return "{" + (int) x + "," + (int) y + "}";
+        return "{" + x + "," + y + "}";
     }
 }
